Fixed: Dasherize tests (#7)
* Fixed: Remove uppercase test

This PR removes the uppercase text test.

The dasherize function was previously modified to not put the leading dash in the case of a capital letter starting the string, so this test is no longer needed. Any other cases are covered by existing tests.

* Fixed: Dasherize tests

Breaks dasherize tests out into two separate ones. One checks that there is no whitespace in the resulting string. The second checks that there are no consecutive dashes in the resulting string.

This replaces a previous test that was failing on bad input.